camera_app_project/backend_requirements.txt
Django==4.2.8
djangorestframework==3.14.0
django-cors-headers==4.3.0
opencv-python==4.8.1.78
Pillow==10.1.0
gunicorn==21.2.0